+++ src/Products/PageTemplates/unicodeconflictresolver.py
@@ -15,7 +15,7 @@
 
 import sys
 
-import ZPublisher
+import ZPublisher.HTTPRequest
 from Acquisition import aq_get
 from Products.PageTemplates.interfaces import IUnicodeEncodingConflictResolver
 from zope.i18n.interfaces import IUserPreferredCharsets